# Estimate performance measurement inaccuracy in tools/perf.sh and tools/run-perf-test.sh