--- a/tools/run-perf-test.sh
+++ b/tools/run-perf-test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# Copyright 2014-2015 Samsung Electronics Co., Ltd.
+# Copyright 2014-2016 Samsung Electronics Co., Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -26,9 +26,9 @@ function exit_err() {
   exit 1
 }
 
-USAGE="Usage:\n    sudo run.sh OLD_ENGINE NEW_ENGINE REPEATS TIMEOUT BENCH_FOLDER"
+USAGE="Usage:\n    sudo run.sh OLD_ENGINE NEW_ENGINE REPEATS TIMEOUT BENCH_FOLDER OUTPUT_FORMAT"
 
-if [ "$#" -ne 5 ]
+if [ "$#" -ne 6 ]
 then
   echo -e "${USAGE}"
   exit_err "Argument number mismatch..."
@@ -39,6 +39,15 @@ ENGINE_NEW="$2"
 REPEATS="$3"
 TIMEOUT="$4"
 BENCH_FOLDER="$5"
+OUTPUT_FORMAT="$6"
+
+if [ "${OUTPUT_FORMAT}" != "-m" ]
+then
+  if [ "${OUTPUT_FORMAT}" != "-c" ]
+  then
+    exit_err "Please, use -m or -c as output format specifier"
+  fi
+fi
 
 if [ "${REPEATS}" -lt 1 ]
 then
@@ -54,13 +63,27 @@ perf_n=0
 mem_n=0
 
 perf_rel_mult=1.0
+perf_rel_inaccuracy_tmp=0
 mem_rel_mult=1.0
+mem_rel_inaccuracy_tmp="-1"
+
+# Unicode "figure space" character
+FIGURE_SPACE=$(echo -e -n "\xE2\x80\x87")
+
+# Unicode "approximately equal" character
+APPROXIMATELY_EQUAL=$(echo -n -e "\xE2\x89\x88")
 
 function run-compare()
 {
   COMMAND=$1
   PRE=$2
   TEST=$3
+  PRECISION=$4
+  UNIT=$5
+
+  ABS_FP_FMT="%$((PRECISION + 4)).$((PRECISION))f$UNIT"
+  REL_FP_FMT="%0.3f"
+  REL_SHOW_PLUS_SIGN_FP_FMT="%+0.3f"
 
   OLD=$(timeout "${TIMEOUT}" ${COMMAND} "${ENGINE_OLD}" "${TEST}") || return 1
   NEW=$(timeout "${TIMEOUT}" ${COMMAND} "${ENGINE_NEW}" "${TEST}") || return 1
@@ -68,20 +91,119 @@ function run-compare()
   #check result
   ! $OLD || ! $NEW || return 1
 
+  OLD_value=$(echo "$OLD " | cut -d ' ' -f 1)
+  OLD_inaccuracy=$(echo "$OLD " | cut -d ' ' -f 2)
+
+  NEW_value=$(echo "$NEW " | cut -d ' ' -f 1)
+  NEW_inaccuracy=$(echo "$NEW " | cut -d ' ' -f 2)
+
   #calc relative speedup
-  rel=$(echo "${OLD}" "${NEW}" | awk '{print $2 / $1; }')
+  eval "rel_mult=\$${PRE}_rel_mult"
+
+  rel=$(echo "${OLD_value}" "${NEW_value}" | awk '{ print $2 / $1; }')
 
   #increment n
   ((${PRE}_n++))
 
-  #accumulate relative speedup
-  eval "rel_mult=\$${PRE}_rel_mult"
+  #calc percent to display
+  PERCENT=$(echo "$rel" | awk '{print (1.0 - $1) * 100; }')
+
+  if [[ "$OLD_inaccuracy" != "" && "$NEW_inaccuracy" != "" ]]
+  then
+    DIFF=$(printf "$ABS_FP_FMT -> $ABS_FP_FMT" $OLD_value $NEW_value)
+    rel_inaccuracy=$(echo "$OLD_value $OLD_inaccuracy $NEW_value $NEW_inaccuracy" | \
+                     awk "{
+                            OLD_value=\$1
+                            OLD_inaccuracy=\$2
+                            NEW_value=\$3
+                            NEW_inaccuracy=\$4
+
+                            rel_inaccuracy = (NEW_value / OLD_value) * sqrt ((OLD_inaccuracy / OLD_value) ^ 2 + (NEW_inaccuracy / NEW_value) ^ 2)
+                            if (rel_inaccuracy < 0) {
+                              rel_inaccuracy = -rel_inaccuracy
+                            }
+
+                            print rel_inaccuracy
+                          }")
+    PERCENT_inaccuracy=$(echo "$rel_inaccuracy" | awk '{ print $1 * 100.0 }')
+
+    ext=$(echo "$PERCENT $PERCENT_inaccuracy" | \
+                awk "{
+                       PERCENT=\$1
+                       PERCENT_inaccuracy=\$2
+
+                       if (PERCENT > 0.0 && PERCENT > PERCENT_inaccuracy) {
+                         print \"[+]\"
+                       } else if (PERCENT < 0 && -PERCENT > PERCENT_inaccuracy) {
+                         print \"[-]\"
+                       } else {
+                         print \"[$APPROXIMATELY_EQUAL]\"
+                       }
+                     }")
+
+    if [[ $rel_inaccuracy_tmp -lt 0 ]]
+    then
+      return 1
+    fi
+
+    eval "rel_inaccuracy_tmp=\$${PRE}_rel_inaccuracy_tmp"
+
+    rel_inaccuracy_tmp=$(echo "$rel $rel_inaccuracy $rel_inaccuracy_tmp" | \
+                         awk "{
+                                rel=\$1
+                                rel_inaccuracy=\$2
+                                rel_inaccuracy_tmp=\$3
+                                print rel_inaccuracy_tmp + (rel_inaccuracy / rel) ^ 2
+                              }")
+
+    eval "${PRE}_rel_inaccuracy_tmp=\$rel_inaccuracy_tmp"
+
+    PERCENT=$(printf "%8s %11s" $(printf "$REL_SHOW_PLUS_SIGN_FP_FMT%%" $PERCENT) $(printf "(+-$REL_FP_FMT%%)" $PERCENT_inaccuracy))
+    PERCENT="$PERCENT : $ext"
+
+    if [ "${OUTPUT_FORMAT}" == "-m" ]
+    then
+      WIDTH=42
+      DIFF=$(printf "%s%s" "$DIFF" "$(printf "%$(($WIDTH - ${#DIFF}))s")")
+      PERCENT=$(printf "%s%s" "$(printf "%$(($WIDTH - ${#PERCENT}))s")" "$PERCENT")
+
+      format="\`%s\`<br>\`%s\`"
+    else
+      format="%20s : %19s"
+    fi
+  else
+    ext=""
+
+    if [[ "$OLD_inaccuracy" != "" || "$NEW_inaccuracy" != "" ]]
+    then
+      return 1;
+    fi
+
+    DIFF=$(printf "$ABS_FP_FMT -> $ABS_FP_FMT" $OLD_value $NEW_value)
+    PERCENT=$(printf "$REL_SHOW_PLUS_SIGN_FP_FMT%%" $PERCENT)
+
+    if [ "${OUTPUT_FORMAT}" == "-m" ]
+    then
+      WIDTH=20
+      DIFF=$(printf "%s%s" "$DIFF" "$(printf "%$(($WIDTH - ${#DIFF}))s")")
+      PERCENT=$(printf "%s%s" "$(printf "%$(($WIDTH - ${#PERCENT}))s")" "$PERCENT")
+
+      format="\`%s\`<br>\`%s\`"
+    else
+      format="%14s : %8s"
+    fi
+  fi
+
   rel_mult=$(echo "$rel_mult" "$rel" | awk '{print $1 * $2;}')
+
   eval "${PRE}_rel_mult=\$rel_mult"
 
-  #calc percent to display
-  percent=$(echo "$rel" | awk '{print (1.0 - $1) * 100; }')
-  printf "%28s" "$(printf "%6s->%6s (%3.3f)" "$OLD" "$NEW" "$percent")"
+  if [ "${OUTPUT_FORMAT}" == "-m" ]
+  then
+    printf "$format" "$DIFF" "$PERCENT" | sed "s/ /$FIGURE_SPACE/g"
+  else
+    printf "$format" "$DIFF" "$PERCENT"
+  fi
 }
 
 function run-test()
@@ -89,11 +211,15 @@ function run-test()
   TEST=$1
 
   # print only filename
-  printf "%40s | " "${TEST##*/}"
-  run-compare "./tools/rss-measure.sh"      "mem"   "${TEST}" || return 1
+  if [ "${OUTPUT_FORMAT}" == "-m" ]
+  then
+    printf "%s | " "${TEST##*/}"
+  else
+    printf "%50s | " "${TEST##*/}"
+  fi
+  run-compare "./tools/rss-measure.sh"      "mem"   "${TEST}" 0 k || return 1
   printf " | "
-  run-compare "./tools/perf.sh ${REPEATS}"  "perf"  "${TEST}" || return 1
-  printf " | "
+  run-compare "./tools/perf.sh ${REPEATS}"  "perf"  "${TEST}" 3 s || return 1
   printf "\n"
 }
 
@@ -108,16 +234,65 @@ function run-suite()
 }
 
 date
-printf "%40s | %28s | %28s |\n" "Benchmark" "RSS<br>(+ is better)" "Perf<br>(+ is better)"
-printf "%40s | %28s | %28s |\n" "---------" "---" "----"
+
+if [ "${OUTPUT_FORMAT}" == "-m" ]
+then
+  echo "Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)"
+  echo "---------: | --------- | ---------"
+else
+  printf "%50s | %25s | %35s\n" "Benchmark" "RSS<br>(+ is better)" "Perf<br>(+ is better)"
+fi
 
 run-suite "${BENCH_FOLDER}"
 
 mem_rel_gmean=$(echo "$mem_rel_mult" "$mem_n" | awk '{print $1 ^ (1.0 / $2);}')
 mem_percent_gmean=$(echo "$mem_rel_gmean" | awk '{print (1.0 - $1) * 100;}')
+if [[ $mem_rel_inaccuracy_tmp != "-1" ]]
+then
+  exit_err "Incorrect inaccuracy calculation for memory consumption geometric mean"
+fi
 
 perf_rel_gmean=$(echo "$perf_rel_mult" "$perf_n" | awk '{print $1 ^ (1.0 / $2);}')
 perf_percent_gmean=$(echo "$perf_rel_gmean" | awk '{print (1.0 - $1) * 100;}')
+if [[ "$perf_rel_inaccuracy_tmp" == "-1" ]]
+then
+  exit_err "Incorrect inaccuracy calculation for performance geometric mean"
+else
+  perf_percent_inaccuracy=$(echo "$perf_rel_gmean $perf_rel_inaccuracy_tmp $perf_n" | \
+                            awk "{
+                                   perf_rel_gmean=\$1
+                                   perf_rel_inaccuracy_tmp=\$2
+                                   perf_n=\$3
 
-printf "%40s | %28s | %28s |\n" "Geometric mean:" "RSS reduction: $mem_percent_gmean%" "Speed up: $perf_percent_gmean%"
+                                   print 100.0 * (perf_rel_gmean ^ (1.0 / perf_n) * sqrt (perf_rel_inaccuracy_tmp) / perf_n)
+                                 }")
+  perf_ext=$(echo "$perf_percent_gmean $perf_percent_inaccuracy" | \
+             awk "{
+                    perf_percent_gmean=\$1
+                    perf_percent_inaccuracy=\$2
+
+                    if (perf_percent_gmean > 0.0 && perf_percent_gmean > perf_percent_inaccuracy) {
+                      print \"[+]\"
+                    } else if (perf_percent_gmean < 0 && -perf_percent_gmean > perf_percent_inaccuracy) {
+                      print \"[-]\"
+                    } else {
+                      print \"[$APPROXIMATELY_EQUAL]\"
+                    }
+                  }")
+  perf_percent_inaccuracy=$(printf "(+-%0.3f%%) : $perf_ext" $perf_percent_inaccuracy)
+fi
+
+gmean_label_text="Geometric mean:"
+
+if [ "${OUTPUT_FORMAT}" == "-m" ]
+then
+  mem_percent_gmean_text=$(printf "RSS reduction: \`%0.3f%%\`" "$mem_percent_gmean")
+  perf_percent_gmean_text=$(printf "Speed up: \`%0.3f%% %s\`" "$perf_percent_gmean" "$perf_percent_inaccuracy")
+  printf "%s | %s | %s\n" "$gmean_label_text" "$mem_percent_gmean_text" "$perf_percent_gmean_text"
+else
+  mem_percent_gmean_text=$(printf "RSS reduction: %0.3f%%" "$mem_percent_gmean")
+  perf_percent_gmean_text=$(printf "Speed up: %0.3f%% %s" "$perf_percent_gmean" "$perf_percent_inaccuracy")
+  printf "%50s | %25s | %51s\n" "$gmean_label_text" "$mem_percent_gmean_text" "$perf_percent_gmean_text"
+fi
+
 date