The new version of measurements table looks like the following:

- same version of engine, 3 measurement iterations

Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)
---------: | --------- | ---------
3d-cube.js | ` 176k ->  176k      `<br>`             +0.000%` | `  6.639s ->   6.631s                      `<br>`                  +0.120% (+-1.972%) : [≈]` | 
bitops-bits-in-byte.js | `  36k ->   36k      `<br>`             +0.000%` | `  7.588s ->   7.604s                      `<br>`                  -0.211% (+-2.736%) : [≈]` | 
bitops-bitwise-and.js | `  28k ->   28k      `<br>`             +0.000%` | `  5.219s ->   5.200s                      `<br>`                  +0.358% (+-4.477%) : [≈]` | 
math-cordic.js | `  52k ->   52k      `<br>`             +0.000%` | `  8.931s ->   8.896s                      `<br>`                  +0.388% (+-6.232%) : [≈]` | 
Geometric mean: | RSS reduction: `0.000%` | Speed up: `0.164% (+-2.101%) : [≈]`

- different versions of engine, 3 measurement iterations

Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)
---------: | --------- | ---------
3d-cube.js | ` 176k ->  172k      `<br>`             +2.273%` | `  6.617s ->   5.885s                      `<br>`                 +11.062% (+-3.753%) : [+]` | 
bitops-bits-in-byte.js | `  36k ->   36k      `<br>`             +0.000%` | `  7.595s ->   6.711s                      `<br>`                 +11.640% (+-3.426%) : [+]` | 
bitops-bitwise-and.js | `  28k ->   28k      `<br>`             +0.000%` | `  5.219s ->   4.235s                      `<br>`                 +18.855% (+-2.898%) : [+]` | 
math-cordic.js | `  52k ->   48k      `<br>`             +7.692%` | `  8.871s ->   7.513s                      `<br>`                 +15.301% (+-3.438%) : [+]` | 
Geometric mean: | RSS reduction: `2.543%` | Speed up: `14.273% (+-1.895%) : [+]`

- same version of engine, 10 measurement iterations

Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)
---------: | --------- | ---------
3d-cube.js | ` 176k ->  176k      `<br>`             +0.000%` | `  6.620s ->   6.621s                      `<br>`                  -0.020% (+-0.714%) : [≈]` | 
bitops-bits-in-byte.js | `  36k ->   36k      `<br>`             +0.000%` | `  7.591s ->   7.580s                      `<br>`                  +0.141% (+-0.636%) : [≈]` | 
bitops-bitwise-and.js | `  28k ->   28k      `<br>`             +0.000%` | `  5.201s ->   5.214s                      `<br>`                  -0.256% (+-0.756%) : [≈]` | 
math-cordic.js | `  52k ->   52k      `<br>`             +0.000%` | `  8.881s ->   8.884s                      `<br>`                  -0.030% (+-0.682%) : [≈]` | 
Geometric mean: | RSS reduction: `0.000%` | Speed up: `-0.041% (+-0.349%) : [≈]`

- different versions of engine, 10 measurement iterations

Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)
---------: | --------- | ---------
3d-cube.js | ` 176k ->  172k      `<br>`             +2.273%` | `  6.633s ->   5.881s                      `<br>`                 +11.338% (+-0.694%) : [+]` | 
bitops-bits-in-byte.js | `  36k ->   36k      `<br>`             +0.000%` | `  7.581s ->   6.711s                      `<br>`                 +11.468% (+-0.526%) : [+]` | 
bitops-bitwise-and.js | `  28k ->   28k      `<br>`             +0.000%` | `  5.209s ->   4.238s                      `<br>`                 +18.646% (+-0.730%) : [+]` | 
math-cordic.js | `  52k ->   48k      `<br>`             +7.692%` | `  8.882s ->   7.505s                      `<br>`                 +15.507% (+-0.692%) : [+]` | 
Geometric mean: | RSS reduction: `2.543%` | Speed up: `14.294% (+-0.376%) : [+]`