@ruben-ayrapetyan 2016-01-20
Perhaps, its better to reduce width of table, so that it would fit into comment field's width.

We could do it by removing spaces. However, this would lead to unaligned sub-columns and so the version with removed spaces would also be less readable than it could be.
The other way is to remove some information or reduce estimated maximum spaces count for some fields.

Please, share your opinion about this.