@zherczeg 2016-01-20
Just a question: RSS is based on random sampling, why don't we simply use the memstats results to track memory progress?