@ruben-ayrapetyan 2016-01-20
@zherczeg, there are plans to add mem-stats info to the table too. However, mem-stats doesn't account stack usage, for example. So, mem-stats info only is not enough.