@ruben-ayrapetyan 2016-01-20
What do you think of the following version? Is it better than the current version in the header of the pull request?

Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)
---------: | --------- | ---------
3d-cube.js | `+2.273% : 176k ->  172k` | `+11.531% (+-0.705%) : [+]`<br>`  6.637s ->   5.872s` | 
bitops-bits-in-byte.js | `+0.000% :  36k ->   36k` | `+11.375% (+-0.598%) : [+]`<br>`   7.584s ->   6.721s` | 
bitops-bitwise-and.js | `+0.000% :  28k ->   28k` | `+19.387% (+-6.461%) : [+]`<br>`   5.261s ->   4.241s` | 
math-cordic.js | `+7.692% :  52k ->   48k` | `+15.925% (+-2.326%) : [+]`<br>`   8.925s ->   7.504s` | 
Geometric mean: | RSS reduction: 2.5429% | Speed up: 14.62% (+-1.833%) [+]