@zherczeg 2016-01-20
I would put the from->to first, and the gain(deviation) after. If you split it inot two columns, I would align the former to left and the latter to right. We don't need to squeeze this to a very small space.