@ruben-ayrapetyan 2016-01-20
Benchmark | RSS<br>(+ is better) | Perf (absolute) | Perf (relative)
---------: | --------- | --------- | ---------
3d-cube.js | `+2.273% : 176k ->  172k` | `  6.637s ->   5.872s` | `+11.531% (+-0.705%) : [+]` |
bitops-bits-in-byte.js | `+0.000% :  36k ->   36k` | `  7.584s ->   6.721s` | `+11.375% (+-0.598%) : [+]` |
bitops-bitwise-and.js | `+0.000% :  28k ->   28k` | `  5.261s ->   4.241s` | `+19.387% (+-6.461%) : [+]` |
math-cordic.js | `+7.692% :  52k ->   48k` | `  8.925s ->   7.504s` | `+15.925% (+-2.326%) : [+]` |
Geometric mean: | RSS reduction: 2.5429% | |  Speed up: 14.62% (+-1.833%) [+]