@zherczeg 2016-01-20
Oh, sorry, I wasn't clear enough.

Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)
---------: | ---------: | ---------:
3d-cube.js | `176k -> 172k (left aligned)`<br>`(right aligned) +2.273% (+-)` | `6.637s -> 5.872s (left aligned) [+]`<br>`   (right aligned) +11.531% (+-0.705%)` | 

So the "two" columns are placed in one "column", but the top values are aligned left, while the bottom values aligned right.