@zherczeg 2016-01-20
Of course they cvan be overlap at some level, but not fully (have some space at left and right).