@ruben-ayrapetyan 2016-01-20
@zherczeg, I see. So, it would be like the following, right?

Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)
---------: | ---------: | ---------:
3d-cube.js | `176k -> 172k         `<br>`          +2.273% (+-)` | `6.637s -> 5.872s [+]                `<br>`                   +11.531% (+-0.705%)` | 
3d-cube.js | `176k -> 172k         `<br>`          +2.273% (+-)` | `6.637s -> 5.872s [+]                `<br>`                   +11.531% (+-0.705%)` |