@zherczeg 2016-01-20
Exactly. I think it is clear enough, but still compact. Anyway, this is just my two cents, I am open to other suggestions.