@ruben-ayrapetyan 2016-01-20
@zherczeg, I think your idea is great!

I've updated pull request. Now, table is generated as the following:

Benchmark | RSS<br>(+ is better) | Perf<br>(+ is better)
---------: | --------- | ---------
3d-cube.js | ` 124k ->  124k      `<br>`             +0.000%` | `  0.319s ->   0.310s                  `<br>`              +2.823% (+-1.259%) : [+]` | 
access-binary-trees.js | `  68k ->   68k      `<br>`             +0.000%` | `  0.179s ->   0.177s                  `<br>`              +0.930% (+-0.185%) : [+]` | 
bitops-bitwise-and.js | `  32k ->   32k      `<br>`             +0.000%` | `  0.308s ->   0.319s                  `<br>`              -3.794% (+-0.818%) : [-]` | 
Geometric mean: | RSS reduction: `0.000%` | Speed up: `0.025% (+-0.756%) : [≈]`