@bzsolt 2016-01-27
@ruben-ayrapetyan, can we support a more console-friendly output format too (like the current one)?