@ruben-ayrapetyan 2016-01-27
@bzsolt, thank you for your comment. I'll try to update this pull request on the next week to add switch for choosing between console and markdown formats.