@ruben-ayrapetyan 2016-02-04
@bzsolt, I've updated the pull request. The update introduced `-c` / `-m` switch that should be passed as sixth command line argument.

Could you, please, check the new version?