@bzsolt 2016-02-05
@ruben-ayrapetyan, I've a question related to 'usage message'. 
Why does it suggest to run with sudo? And I think there is a typo, the ```run.sh``` will be ```run-perf-test.sh```
Apart from minor things above, it looks good. Thanks for the modification.